fix: check nftstorage.link status in any-url path
- check which gateway we should redirect you to
- fix bugs

License: (Apache-2.0 AND MIT)
