+++ src/any-url.js
@@ -1,4 +1,4 @@
-import { InvalidUrlError } from './errors.js'
+import { pickGatewayHost } from './utils/status-check.js'
 import { findPathUrl } from './utils/cid'
 
 /**
@@ -8,6 +8,7 @@ import { findPathUrl } from './utils/cid'
  * @param {import('./env').Env} env
  */
 export async function anyUrlGet (request, env) {
-  const url = findPathUrl(new URL(request.url), env.IPFS_GATEWAY_HOSTNAME)
+  const hostname = await pickGatewayHost(caches.default, env)
+  const url = findPathUrl(new URL(request.url), hostname)
   return Response.redirect(url, 302)  
 }
